# frontend/schema/openapi.json
{
  "components": {
    "schemas": {
      "ActionResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameStateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "ActionResponse",
        "type": "object"
      },
...
      "GameStateResponse": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "current_set": {
            "title": "Current Set",
            "type": "integer"
          },
          "match_finished": {
            "title": "Match Finished",
            "type": "boolean"
          },
          "serve": {
            "title": "Serve",
            "type": "string"
          },
          "simple_mode": {
            "title": "Simple Mode",
            "type": "boolean"
          },
          "team_1": {
            "$ref": "#/components/schemas/TeamState"
          },
          "team_2": {
            "$ref": "#/components/schemas/TeamState"
          },
          "visible": {
            "title": "Visible",
            "type": "boolean"
          }
        },
        "required": [
          "current_set",
          "visible",
          "simple_mode",
          "match_finished",
          "team_1",
          "team_2",
          "serve",
          "config"
        ],
        "title": "GameStateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InitRequest": {
        "properties": {
          "oid": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Oid",
            "type": "string"
          },
          "output_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Url"
          },
          "points_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points Limit"
          },
          "points_limit_last_set": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points Limit Last Set"
          },
          "sets_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sets Limit"
          }
        },
        "required": [
          "oid"
        ],
        "title": "InitRequest",
        "type": "object"
      },
...
      "TeamState": {
        "properties": {
          "scores": {
            "additionalProperties": true,
            "title": "Scores",
            "type": "object"
          },
          "serving": {
            "title": "Serving",
            "type": "boolean"
          },
          "sets": {
            "title": "Sets",
            "type": "integer"
          },
          "timeouts": {
            "title": "Timeouts",
            "type": "integer"
          }
        },
        "required": [
          "sets",
          "timeouts",
          "scores",
          "serving"
        ],
        "title": "TeamState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Volley Overlay Control",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session/init": {
      "post": {
        "description": "Initialise (or re-use) a game session for the given overlay ID.",
        "operationId": "init_session_api_v1_session_init_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Init Session",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
...
  }
}